requirements.txt
asgiref==3.5.0
autopep8==1.6.0
boto3==1.23.10
botocore==1.26.10
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
jmespath==1.0.0
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
